simplify memchr

+++ srcs/memory/ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: rpinoit <rpinoit@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/10/29 20:38:34 by rpinoit           #+#    #+#             */
-/*   Updated: 2019/04/13 23:02:07 by rpinoit          ###   ########.fr       */
+/*   Updated: 2019/04/14 16:44:13 by rpinoit          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -59,7 +59,7 @@ inline static void	set_bytes(unsigned char **pdst, unsigned char c, size_t bytes
 	}
 }
 
-void			*ft_memset(void *dst, int c, size_t len)
+void				*ft_memset(void *dst, int c, size_t len)
 {
 	unsigned long	cccc;
 
@@ -69,7 +69,7 @@ void			*ft_memset(void *dst, int c, size_t len)
 		cccc |= cccc << 8;
 		cccc |= cccc << 16;
 		cccc |= cccc << 32;
-		align_word((unsigned char **)&dst, (unsigned char)c, (unsigned long)dst % MEM_WORD_LEN);
+		align_word((unsigned char **)&dst, (unsigned char)c, (size_t)dst % MEM_WORD_LEN);
 		len -= (unsigned long)dst % MEM_WORD_LEN;
 		set_blocks((unsigned long **)&dst, cccc, len / MEM_BLOCK_SIZE);
 		len %= MEM_BLOCK_SIZE;
